+++ tests/test_cli_test.py
@@ -1,44 +1,30 @@
 """Test the nbreport test CLI (nbreport.cli.test).
 """
 
-from pathlib import Path
-
-from click.testing import CliRunner
-
 import nbreport.cli.main
 from nbreport.instance import ReportInstance
 
 
-def test_basic(tmpdir):
+def test_basic(testr_000_path, runner):
     """Test with no arguments except repo path.
     """
-    runner = CliRunner()
-
-    repo_path = Path(__file__).parent / 'TESTR-000'
-    repo_path = repo_path.resolve()
-
     with runner.isolated_filesystem():
         args = [
             'test',  # subcommand
-            str(repo_path),  # first argument
+            str(testr_000_path),  # first argument
         ]
         result = runner.invoke(nbreport.cli.main.main, args)
 
         assert result.exit_code == 0
 
 
-def test_basic_with_dirname_arg(tmpdir):
+def test_basic_with_dirname_arg(testr_000_path, runner):
     """Test with --id and -d arguments.
     """
-    runner = CliRunner()
-
-    repo_path = Path(__file__).parent / 'TESTR-000'
-    repo_path = repo_path.resolve()
-
     with runner.isolated_filesystem():
         args = [
             'test',  # subcommand
-            str(repo_path),  # first argument
+            str(testr_000_path),  # first argument
             '--id', '1',
             '-d', 'TESTR-000-1',
         ]
@@ -47,20 +33,14 @@ def test_basic_with_dirname_arg(tmpdir):
         assert result.exit_code == 0
 
 
-def test_config_option(tmpdir):
+def test_config_option(testr_000_path, runner):
     """Test with config (-c) options to configure the template rendering.
     """
-    runner = CliRunner()
-
-    repo_path = Path(__file__).parent / 'TESTR-000'
-    repo_path = repo_path.resolve()
-
     with runner.isolated_filesystem():
         args = [
             '--log-level', 'debug',
             'test',  # subcommand
-            str(repo_path),  # first argument
-            '-c', 'title', 'My sick report',
+            str(testr_000_path),  # first argument
             '-c', 'a', '100',
             '-c', 'b', '200',
         ]
@@ -73,8 +53,8 @@ def test_config_option(tmpdir):
         nb = instance.open_notebook()
 
         assert nb.cells[0].source == (
-            "# My sick report\n"
-            "\n"
+            "**TESTR-000-test**\n\n"
+            "# Test Report\n\n"
             "- By: Test Bot\n"
             "- Date: 2018-07-18"
         )
@@ -84,11 +64,9 @@ def test_config_option(tmpdir):
         )
 
 
-def test_from_git_clone(tmpdir):
+def test_from_git_clone(runner):
     """Test creating an instance from a GitHub original repository.
     """
-    runner = CliRunner()
-
     with runner.isolated_filesystem():
         args = [
             '--log-level', 'debug',
@@ -96,7 +74,6 @@ def test_from_git_clone(tmpdir):
             'https://github.com/lsst-sqre/nbreport',
             '--git-ref', 'master',
             '--git-subdir', 'tests/TESTR-000',
-            '-c', 'title', 'My sick report',
             '-c', 'a', '100',
             '-c', 'b', '200',
         ]