+++ tests/test_templating.py
@@ -1,8 +1,6 @@
 """Tests for the nbreport.templating module.
 """
 
-from pathlib import Path
-
 import nbformat
 import pytest
 
@@ -69,22 +67,20 @@ def test_render_notebook():
     'use_pathlib',
     [(True,), (False,)]
 )
-def test_load_template_environment_from_fs(use_pathlib):
+def test_load_template_environment_from_fs(use_pathlib, testr_000_path):
     """Test loading the template environment in tests/TESTR-000 using the
     ``cookiecutter.json`` file as context.
 
     Test passing the context file's path as both a pathlib.Path and as a
     string.
     """
-    base_dir = Path(__file__).parent / 'TESTR-000'
-    context_path = base_dir / 'cookiecutter.json'
+    context_path = testr_000_path / 'cookiecutter.json'
 
     if not use_pathlib:
         context_path = str(context_path)
 
     context, jinja_env = load_template_environment(context_path=context_path)
 
-    assert context['cookiecutter']['title'] == 'Test Report'
     assert context['cookiecutter']['username'] == 'Test Bot'
     assert context['cookiecutter']['generated_iso8601'] == '2018-07-18'
     assert context['cookiecutter']['a'] == 10
@@ -96,32 +92,3 @@ def test_load_empty_template_environment():
     """
     context, jinja_env = load_template_environment()
     assert len(context['cookiecutter'].keys()) == 0
-
-
-def test_render_ipynb():
-    """Proof-of-concept for rendering the templated ipynb notebook in
-    ``tests/TESTR-000/``.
-
-    This test exercises obtaining a context from the cookiecutter.json file,
-    and rendering a full ipynb given that context.
-    """
-    base_dir = Path(__file__).parent / 'TESTR-000'
-    context_path = base_dir / 'cookiecutter.json'
-    ipynb_path = base_dir / 'TESTR-000.ipynb'
-
-    context, jinja_env = load_template_environment(context_path=context_path)
-    notebook = nbformat.read(str(ipynb_path.resolve()),
-                             as_version=nbformat.NO_CONVERT)
-
-    notebook = render_notebook(notebook, context, jinja_env)
-
-    assert notebook.cells[0].source == (
-        '# Test Report\n'
-        '\n'
-        '- By: Test Bot\n'
-        '- Date: 2018-07-18'
-    )
-    assert notebook.cells[1].source == (
-        "answer = 10 + 32\n"
-        "print('The answer is {}'.format(answer))"
-    )